Update landing page design
